Removal of gmd location
I deleted the addition of gmd to the schema location in the "xsl:stylesheet" attribute. It breaks validation of records external to geonetwork once exported. Although not necessary for Gemini, it is useful for validation. There maybe a good reason you have this in though. I currently have to remove it on the way out.